+++ algorithm/cgal/Remesh.h
@@ -4,18 +4,18 @@
 
 static int Remesh(Geometry* geometry, size_t count, size_t iterations,
                   size_t relaxation_steps, double target_edge_length) {
-  int size = geometry->getSize();
+  size_t size = geometry->getSize();
 
   geometry->copyInputMeshesToOutputMeshes();
   geometry->transformToAbsoluteFrame();
 
   std::vector<const Surface_mesh*> selections;
 
-  for (int selection = count; selection < size; selection++) {
+  for (size_t selection = count; selection < size; selection++) {
     selections.push_back(&geometry->mesh(selection));
   }
 
-  for (int nth = 0; nth < count; nth++) {
+  for (size_t nth = 0; nth < count; nth++) {
     remesh<Kernel>(geometry->mesh(nth), selections, iterations,
                    relaxation_steps, target_edge_length);
   }
